backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.13.1
PyPDF2==3.0.1
python-docx==1.1.0
